pyproject.toml
[project]
name = "itrading"
version = "0.1.0"
description = "早盘量化选股系统 - 基于多维度指标的股票选择器"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv",
    "pandas",
    "numpy", 
    "matplotlib",
    "google-genai",
    "tushare",
    "akshare",
    "qstock",
    "pyfolio",
    "backtrader"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
